+++ carveracontroller/main.py
@@ -90,6 +90,7 @@ from kivy.properties import BooleanProperty
 from kivy.graphics import Color, Rectangle, Ellipse, Line
 from kivy.properties import ObjectProperty, NumericProperty, ListProperty
 from kivy.config import Config
+from kivy.metrics import Metrics
 
 from serial.tools.list_ports import comports
 from functools import partial
@@ -551,7 +552,10 @@ class MakeraConfigPanel(SettingsWithSidebar):
     def on_config_change(self, config, section, key, value):
         app = App.get_running_app()
         if not app.root.config_loading:
-            if section != 'Restore':
+            if section == 'carvera':
+                app.root.controller_setting_change_list[key] = value
+                app.root.config_popup.btn_apply.disabled = False
+            elif section != 'Restore':
                 app.root.setting_change_list[key] = Utils.to_config(app.root.setting_type_list[key], value).strip()
                 app.root.config_popup.btn_apply.disabled = False
             elif key == 'restore' and value == 'RESTORE':
@@ -1360,6 +1364,8 @@ class Makera(RelativeLayout):
         self.setting_list = {}
         self.setting_type_list = {}
         self.setting_default_list = {}
+        self.controller_setting_change_list = {}
+        self.load_controller_config()
 
         self.usb_event = lambda instance, x: self.openUSB(x)
         self.wifi_event = lambda instance, x: self.openWIFI(x)
@@ -1415,8 +1421,26 @@ class Makera(RelativeLayout):
                 app.root.controller.jog("Z{}".format(app.root.step_z.text))
             elif key == 281:  # page down
                 app.root.controller.jog("Z-{}".format(app.root.step_z.text))
-        
     
+    def load_controller_config(self):
+        config_def_file = os.path.join(os.path.dirname(__file__),'controller_config.json')
+        with open(config_def_file) as file:
+            controller_config_definition = json.load(file)
+        controller_config = []
+
+        # Set default controller config values
+        for setting in controller_config_definition:
+            if 'default' in setting:
+                Config.setdefault('carvera', setting['key'], setting['default'])
+                setting.pop('default', None)
+            controller_config.append(setting)
+
+        self.config_popup.settings_panel.add_json_panel('Controller', Config, data=json.dumps(controller_config))
+
+    def apply_controller_settings(self):
+        pass
+
+
     def open_download(self):
         webbrowser.open(DOWNLOAD_ADDRESS, new = 2)
     
@@ -3330,21 +3354,36 @@ class Makera(RelativeLayout):
                         advanced.pop('section')
                     elif 'default' in advanced:
                         advanced.pop('default')
-
-                self.config_popup.settings_panel.add_json_panel('Basic', self.config, data=json.dumps(basic_config))
-                self.config_popup.settings_panel.add_json_panel('Advanced', self.config, data=json.dumps(advanced_config))
+                self.config_popup.settings_panel.add_json_panel('Machine - Basic', self.config, data=json.dumps(basic_config))
+                self.config_popup.settings_panel.add_json_panel('Machine - Advanced', self.config, data=json.dumps(advanced_config))
                 self.config_popup.settings_panel.add_json_panel('Restore', self.config, data=json.dumps(restore_config))
         return True
 
     # -----------------------------------------------------------------------
     def apply_setting_changes(self):
+        if self.setting_change_list:
+            self.apply_machine_setting_changes()
+        if self.controller_setting_change_list:
+            self.apply_controller_setting_changes()
+
+
+    def apply_machine_setting_changes(self):
         for key in self.setting_change_list:
             self.controller.setConfigValue(key, self.setting_change_list[key])
             time.sleep(0.1)
         self.setting_change_list.clear()
         self.config_popup.btn_apply.disabled = True
-        self.message_popup.lb_content.text = tr._('Settings applied, need reset to take effect !')
+        self.message_popup.lb_content.text = tr._('Settings applied, need machine reset to take effect !')
         self.message_popup.open()
+
+    
+    def apply_controller_setting_changes(self):
+        if self.controller_setting_change_list.get("ui_density_override") or self.controller_setting_change_list.get("ui_density"):
+            self.message_popup.lb_content.text = tr._('UI Density changed, restart application to apply.')
+            self.message_popup.open()
+
+        self.config_popup.btn_apply.disabled = True
+
 
     # -----------------------------------------------------------------------
     def open_setting_restore_confirm_popup(self):
@@ -3716,6 +3755,10 @@ def android_tweaks():
     except ImportError:
         print("Pyjnius Import Fail.")
 
+def load_app_configs():
+    if Config.get('carvera', 'ui_density_override') == "1":
+        Metrics.set_density(float(Config.get('carvera', 'ui_density')))
+
 def set_config_defaults(default_lang):
     if not Config.has_section('carvera'):
         Config.add_section('carvera')
@@ -3807,6 +3850,7 @@ def main():
     default_lang = init_lang()
     tr = Lang(default_lang)
     set_config_defaults(default_lang)
+    load_app_configs()
     HALT_REASON = load_halt_translations(tr)
 
     base_path = app_base_path()